도전 과제 - 제네릭 추가 및 double타입의 값 계산

+++ src/com/example/challengeCalculator/OperatorType.java
@@ -2,17 +2,16 @@ package com.example.challengeCalculator;
 
 public enum OperatorType {
     SUM {
-        public int apply(int num1, int num2) {return num1 + num2;}
+        public double apply(double num1, double num2) {return num1 + num2;}
     },
     SUB{
-        public int apply(int num1, int num2) {return num1 - num2;}
+        public double apply(double num1, double num2) {return num1 - num2;}
     },
     MUL {
-        public int apply(int num1, int num2) {return num1 * num2;}
+        public double apply(double num1, double num2) {return num1 * num2;}
     },
     DIV{
-        public int apply(int num1, int num2) {return num1 / num2;}
+        public double apply(double num1, double num2) {return num1 / num2;}
     };
-
-    public abstract int apply(int num1, int num2);
+    public abstract double apply(double num1, double num2);
 }
